Develop

+++ public/javascript/script.js
@@ -22,6 +22,11 @@ const clickedId = (element => {
     }
     return element.dataset.id;
 })
+
+const week_of = () => {
+    return convertDateTo('reverse_date', currentWeekEl.innerText);
+}
+
 // Converts a date object to a formatted date or time string
 const convertDateTo = (format, date) => {
     const date_object = new Date(date);
@@ -87,6 +92,8 @@ const convertDateTo = (format, date) => {
             return dayname[weekday].long;
         case 'weekday_name_short':
             return dayname[weekday].short;
+        default:
+            return false;
     }
 }
 const addDaysTo = (date, days) => {
@@ -107,15 +114,9 @@ const get_weekOf_for = (date) => {
 // returns the sorted event objects for a particular day from the downloaded data
 const events = (date => {
     medium_date = convertDateTo('medium_date', date);
-    days_events = event_items.filter(e => {
+    return event_items.filter(e => {
         return convertDateTo('medium_date', e.date) === medium_date; 
-    })
-    .sort((a,b) => { 
-        if(a.date < b.date){return -1}
-        else if (a.date > b.date) {return 1}
-        return 0
     });
-     return days_events
 });
 const updateClock = () => {
     const now = new Date()
@@ -127,7 +128,7 @@ const updateClock = () => {
 // gets the week's data from the database
 async function load_events() {
     
-    const response = await fetch ("/api/events", {
+    const response = await fetch (`/api/events/weekof/${week_of()}`, {
         method: 'get',
         headers: {'Content-Type':'application/json'}
     })
@@ -409,7 +410,7 @@ function display_modal() {
 // Start of event listens
 
 // change of the weekof date
-document.querySelector('.week-selector').addEventListener('click', event => {
+document.querySelector('.week-controls').addEventListener('click', event => {
 
     let currentWeek = new Date(currentWeekEl.innerText);
 
@@ -511,20 +512,19 @@ document.querySelector('#cancel').addEventListener('click', (e) => {
 document.querySelector('#delete').addEventListener('click', (e) => {
     e.preventDefault();
     const currentWeek = new Date(currentWeekEl.innerText);
+    if(confirm("Delete event?")) {
+        document.querySelector('#event-input').style.display = "none";
     
-    document.querySelector('#event-input').style.display = "none";
-    
-    delete_event(editing_event).then(data => {
-            editing_event = null;
-        })
-        .then(generateCalendar);
+        delete_event(editing_event).then(data => {
+                editing_event = null;
+            })
+            .then(generateCalendar);
+    }
+
 });
 // logout button clicked
 document.querySelector('#logout').addEventListener('click', logout);
 
-setInterval(() => {
-    updateClock();
-    generateCalendar();
-}, 1000)
-
 init();
+
+setInterval(updateClock, 1000);
